requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
eval_type_backport==0.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
mistralai==1.6.0
py-bingx==0.4
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
